Deep merge global and default options with user-supplied options
This way one can override or add arbitrary keys and values to the
`global_options` and `defaults_options` hashes without having to reproduce
the whole hash.

This is done by deep merging the default hashes for `global_options` and
`defaults_options` specified in `haproxy::params` with their pendants
specified as the class parameters. For array values that are present in
both a default hash and a user-supplied hash the arrays are concatenated
and duplicates removed.

Because of the need for deep merging the `deep_merge()` function from
puppetlabs-stdlib 4.2.0 is copied here, renamed `haproxy_deep_merge()`
and extended to support merging array values

Contains updated spec tests and documentation.

Discussed in PR #121.