Use consul-template for balancemembers
This patch adds a second type of resource for
adding balancemembers, namely: haproxy/balancemember_consul.

These balance members use consul-tempalte as a replacement for
exported resources and assume that both consul as well as
consul-template are installed on any systems using this
defined resource type.

Consul-template offers an alternative to exported resources
that allows for much more dynamic updates to configuration
based on the current state of a load balanced service (it
can allow nearly instantaneous updates of config when pool members
or added, removed, or start behaving badly)

Puppet configures the consul templates which rely on consul-template
to actually convert them into the real haproxy file on disk.