MODULES-2951
Added support for individual weights for balancermembers.

Changes done to manifests/balancermember.pp and templates/haproxy_balancermember.erb

In manifests/balancermember.pp I have added in the definition of data type a line for a weights array that is undefined by default.

In templates/haproxy_balancermember.erb I have added a check to see if weights contain data, and if so print the weight for the server line.